Refactor the Monitor code.
